+++ src/mrgsolve-export/expression.js
@@ -30,7 +30,6 @@ Expression.prototype.toCString = function(){
   };
 
   return this.exprParsed
-    //.translate(math.expression.translator.to['dbsolve'])
     .toString({
       parenthesis: 'keep',
       implicit: 'show',